+++ eslint.config.mjs
@@ -39,7 +39,7 @@ export default [
         console: 'readonly',
       },
       parserOptions: {
-        project: true,
+        project: 'tsconfig.spec.json',
         tsconfigRootDir: './',
       },
     },