+++ vitest.setup.ts
@@ -1,5 +1,6 @@
-import './jest.extend';
+import './vitest.extend';
 import { getEnv, setEnv } from './src/env';
+import { beforeAll } from 'vitest';
 
 // set custom env
 beforeAll(() => setEnv({ ...getEnv(), window, document }));