+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [next]
 
+- ci(): Switch from Jest to Vitest [#10420](https://github.com/fabricjs/fabric.js/pull/10420)
 - chore(): Update canvas and jsdom [#10417](https://github.com/fabricjs/fabric.js/pull/10417)
 
 ## [6.6.2]